(Temporarily?) unconditionally enable Object property configurability and enumerability.
As opposed to having a (random) property on a (random) object be
randomly configurable and enumerable. These were causing subtle bugs in
another library in which a random variable returned from this library
was not able to be accessed properly. The exact cause of this bug is
still unknown to me, but (if and when) I figure it out, I'll restore
this functionality and make the corresponding changes in the other
library.